pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "etl-pipeline-mvp"
version = "0.1.0"
description = "Natural Language to ETL Pipeline MVP"
dependencies = [
    "apache-airflow>=2.9.0",
    "duckdb>=0.9.0",
    "pandas>=2.1.0", 
    "pyarrow>=14.0.0",
    "requests>=2.31.0",
    "pydantic>=2.5.0",
    "jinja2>=3.1.0",
    "openai>=1.6.0",
    "python-dotenv>=1.0.0"
]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
select = ["E", "F", "W", "I"]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
